+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ### 0.5.0
 #### Added
 - DateTime functions.
+- ``PAGINATION`` function.
 
 #### Fixed
 - Unable to define variables and make function calls before FILTER, SORT and etc statements.
+- Unable to use params in LIMIT clause
 - ``INNER_HTML`` returns outer HTML instead for dynamic elements.
 - ``INNER_TEXT`` returns HTML instead from dynamic elements.